Services: separate word formatting from dictionary
